requirements.txt
torch
gradio
diffusers
transformers
opencv-python
Pillow
numpy
pandas
